New optimizer version
Crossover operation

The crossover operation mix a melody measure gave as parent with the current individual being evaluated in the loop. There is no defined mutation strength. For every note in the measure, a random value is generated (0 or 1). If true (1), the individual receives a note from the parent. If false (0), the individual maintains the note value.

Mutation operator

For each item in the measure (the individual), a random number is generated and if the number is less than the mutation strength, the value of the item is increased by one.

Evaluate function

The value generated by the evaluate function is to be defined.