NSGA-II parents substitution
The algorithm starts from the last front, removing individuals that won't be in next generation.

If there is a front that will let the ranking with fewer elements than allowed, it cannot be completely removed. Then, the 'distances' vector is created to sort front elements by crowding distance.

Finally, individuals with the lowest crowding distance  value are removed from ranking.